+++ test/unit/data_structures/containers/iqueue.h
@@ -31,7 +31,7 @@ class IqueueTestSuite {
 
   /// Enqueue elements
   void EnqueueTest(size_t count = 30) {
-    for (int i = 0; i < count; ++i) {
+    for (size_t i = 0; i < count; ++i) {
       hipc::OffsetPointer p;
       auto page = alloc_->template
         AllocateConstructObjs<T>(1, p);
@@ -44,7 +44,7 @@ class IqueueTestSuite {
   /// Dequeue and then re-enqueue
   void DequeueTest(size_t count = 30) {
     std::vector<T*> tmp(count);
-    for (int i = 0; i < count; ++i) {
+    for (size_t i = 0; i < count; ++i) {
       tmp[i] = obj_.dequeue();
     }
     for (int i = count - 1; i >= 0; --i) {
@@ -56,7 +56,7 @@ class IqueueTestSuite {
 
   /// Forward iterator
   void ForwardIteratorTest(size_t count = 30) {
-    int fcur = 0;
+    size_t fcur = 0;
     for (T *page : obj_) {
       REQUIRE(page->page_size_ == fcur);
       ++fcur;
@@ -66,7 +66,7 @@ class IqueueTestSuite {
   /// Constant Forward iterator
   void ConstForwardIteratorTest(size_t count = 30) {
     const Container &obj = obj_;
-    int fcur = 0;
+    size_t fcur = 0;
     for (auto iter = obj.cbegin(); iter != obj.cend(); ++iter) {
       T *page = *iter;
       REQUIRE(page->page_size_ == fcur);
@@ -77,9 +77,9 @@ class IqueueTestSuite {
   /// Dequeue an element in the middle of the queue
   void DequeueMiddleTest() {
     size_t count = obj_.size();
-    int mid = obj_.size() / 2;
+    size_t mid = obj_.size() / 2;
     auto iter = obj_.begin();
-    for (int i = 0; i < mid; ++i) {
+    for (size_t i = 0; i < mid; ++i) {
       ++iter;
     }
     auto page = obj_.dequeue(iter);
